fix: floating point calculations issue

+++ src/models/delegator_data.rs
@@ -13,5 +13,5 @@ pub struct DelegatorData {
     pub last_update_block: u64,
     pub epoch_id: String,
     pub rewards: String,
-    pub apy: f64, // New field for APY
+    pub apy: u128, // New field for APY
 }
